Support a server-first mode
This change adds a `--reverse` flag to the client and server.  When
set, the server sends the message and the client echoes it.

The Helm chart has been updated to support this as well. When set, the
server is annotated with a opaque-ports annotation so that the proxy
skips protocol detection.